+++ src/narrator/core/__init__.py
@@ -1,0 +1,19 @@
+"""Simulation core exports."""
+
+from narrator.core.clock import GlobalClock
+from narrator.core.interrupt import InterruptManager, InterruptRule, InterruptSignal
+from narrator.core.rule_engine import Rule, RuleContext, RuleEngine, RuleEngineResult, RuleExecutionRecord
+from narrator.core.seed import SeedManager
+
+__all__ = [
+    "GlobalClock",
+    "InterruptManager",
+    "InterruptRule",
+    "InterruptSignal",
+    "Rule",
+    "RuleContext",
+    "RuleEngine",
+    "RuleEngineResult",
+    "RuleExecutionRecord",
+    "SeedManager",
+]